+++ drink-test/src/main/java/io/github/kiryu1223/drink/test/oracle/DateTimeTest.java
@@ -95,12 +95,11 @@ public class DateTimeTest extends BaseTest
     @Test
     public void dateTimeDiffTest()
     {
-        int one = client.queryEmptyTable()
+        long one = client.queryEmptyTable()
                 .endSelect(() -> dateTimeDiff(SqlTimeUnit.DAY, "1996-10-27", "2000-01-01"))
                 .first();
 
         Assert.assertEquals(1161, one);
-
 
         int one1 = client.queryEmptyTable()
                 .endSelect(() -> dateTimeDiff(SqlTimeUnit.DAY, "1996-10-27", LocalDate.of(2020, 5, 5)))
